Push Tokend connection configs into the TokendClient class.
