+++ lib/transformers/tokend-client.js
@@ -1,17 +1,26 @@
 'use strict';
 
 const http = require('http');
+const Immutable = require('immutable');
 
 class TokendClient {
   /**
    * Create a new client for connecting to Tokend.
    *
-   * @param {String} host  host to connect to Tokend on
-   * @param {Number} port  port to connect to Tokend on
+   * @param {Object} options
+   * @param {String} options.host  host to connect to Tokend on
+   * @param {Number} options.port  port to connect to Tokend on
    */
-  constructor(host, port) {
-    this._host = host;
-    this._port = port;
+  constructor(options) {
+    let defaultOptions = Immutable.Map({
+      host: Config.get('tokend:host'),
+      port: Config.get('tokend:port')
+    });
+
+    defaultOptions = defaultOptions.mergeDeep(options);
+
+    this._host = defaultOptions.get('host');
+    this._port = parseInt(defaultOptions.get('port'), 10);
     this._pendingGetRequests = {};
   }
