+++ lib/transformers/tokend.js
@@ -38,21 +38,11 @@ function collectTransformables(properties, keyPath) {
 }
 
 /**
- * @param {Object}  options
+ * @param {Object} options  See TokendClient for options
  */
 class TokendTransformer {
   constructor(options) {
-    let defaultOptions = Immutable.Map({
-      host: Config.get('tokend:host'),
-      port: Config.get('tokend:port')
-    });
-
-    defaultOptions = defaultOptions.mergeDeep(options);
-
-    const host = defaultOptions.get('host');
-    const port = parseInt(defaultOptions.get('port'), 10);
-
-    this._client = new TokendClient(host, port);
+    this._client = new TokendClient(options);
   }
 
   /**
